InvisiScan_CCTV_Analysis/requirements.txt
asttokens==3.0.0
Brotli==1.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
filelock==3.12.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.2.1
networkx==3.0
numpy==1.24.3
pillow==10.0.0
PySocks==1.7.1
PyYAML==6.0.1
requests==2.31.0
typing_extensions
urllib3==2.0.4
win-inet-pton==1.1.0
colorama==0.4.6
comm==0.2.2
debugpy==1.8.12
decorator==5.1.1
einops==0.8.0
executing==2.2.0
fsspec==2025.2.0
gmpy2==2.1.5
huggingface-hub==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
jupyter_client==8.6.3
opencv-python==4.9.0.80
cvlib==0.2.7
ultralytics==8.2.21
jupyter_core==5.7.2
matplotlib-inline==0.1.7
mkl-service==2.4.2
nest-asyncio==1.6.0
packaging==24.2
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==6.1.1
pure_eval==0.2.3
Pygments==2.19.1
python-dateutil==2.9.0.post0
pyzmq==26.2.1
regex==2024.11.6
safetensors==0.5.2
six==1.17.0
stack-data==0.6.3
sympy==1.13.1
timm==1.0.14
tokenizers==0.21.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
wcwidth==0.2.13
Flask==3.0.3
flask_cors==4.0.0
